fix: ngClass fails on sVG in IE11
This commit is an update for shims_for_IE with the latest version of the
classlist polyfill.

The minified version of the classlist polyfill used was out of date
(https://github.com/eligrey/classList.js/pull/52)

Specifically it was not testing for classlist support on svg elements,
causing ngClass directives on svg elements to fail (and I presume any
other directives with dependency on classlist as well)

note, the PR mentioned above has been accepted by the author.